Create README.md
I have written a description of the purpose of this project, including a description of the pages individually explaining what value they give to the user experience. I have written what is the procedure for the Deployment of the project and what are the Technologies that I have used and for what. I also reported all the external sources used such as Youtube and other websites. I also written what testing I have done to my code before to push my code and publish it and finally I said on what devices I tested this project and using what browser.